cgen: matcher: meta: cast probability to unsigned int, instead of int
Casting the probability to an integer could trigger an overflow.

+++ src/bpfilter/cgen/matcher/meta.c
@@ -69,7 +69,7 @@ _bf_matcher_generate_meta_probability(struct bf_program *program,
     EMIT(program, BPF_EMIT_CALL(BPF_FUNC_get_prandom_u32));
     EMIT_FIXUP_JMP_NEXT_RULE(
         program, BPF_JMP_IMM(BPF_JGT, BPF_REG_0,
-                             (int)(UINT32_MAX * (proba / 100.0)), 0));
+                             (uint32_t)((uint64_t)UINT32_MAX * (proba / 100.0)), 0));
 
     return 0;
 }
